BOT-3 Move MongoCode in directory
